+++ NEWS.md
@@ -1,4 +1,13 @@
-# rhoR 1.2.0.0
+## rhoR 1.2.1.2
+* `getHandSetIndices()` bug fixed for sampling positive indices
+* `createSimulatedCodeSet()` floating point error fixed for exact matching kappa min and max values
+* Implented full test suite for package
+
+## rhoR 1.2.1.1
+
+## rhoR 1.2.1.0
+
+## rhoR 1.2.0.0
 
 * `getHandSetIndices()` added as a new function. See `?rhoR::gethandSetIndices`
 * `createSimulatedCodeSet()` bug fixed when generating the set to match criteria